Fix segfault
in fss_prep_replace(), if one of the replacement pair keys is empty, or if we
cannot add it to the search tree, we set its value in res->replace to null. But
we don't advance the array pointer, like we do in line 164. This leaves the end
of the array neither nulled nor populated, which means that when we iterate on
m_fss_r->replace in lines 38-44, we are potentially calling smart_delete() on
junk RAM. So we have to make sure to initialize any uninitialized slots in the
replacement_array.

Change-Id: Iddfaea242bc24e17e1df161bf45f0aa9a22bf548